put back stream true
